testflask/requirements.txt
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pgpdump3==1.5.2
psycopg2-binary==2.9.4
py-pgp==0.0.1
pycodestyle==2.9.1
pycrypto==2.6.1
python-magic==0.4.27
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
zope.interface==5.5.0
